already create new users in DB
