+++ package.json
@@ -21,12 +21,13 @@
     "react-dom": "^18",
     "react-hook-form": "^7.51.1",
     "react-image-gallery": "^1.3.0",
+    "react-router-dom": "^6.23.0",
+    "react-slick": "^0.30.2",
     "react-stripe-checkout": "^2.6.3",
     "react-youtube": "^10.1.0",
+    "slick-carousel": "^1.8.1",
     "stripe": "^15.1.0",
-    "swiper": "^11.1.0",
-    "react-slick": "^0.30.2",
-    "slick-carousel": "^1.8.1"
+    "swiper": "^11.1.0"
   },
   "devDependencies": {
     "autoprefixer": "^10.0.1",
